feature: Asset Creation
