+++ apps/ai-dial-admin/src/components/Toolsets/View/Properties.tsx
@@ -2,7 +2,7 @@
 
 import { FC } from 'react';
 
-import EntityMainProperties from '@/src/components/EntityMainProperties/EntityMainProperties';
+import DeploymentProperties from '@/src/components/EntityMainProperties/Properties/DeploymentProperties';
 import EntityHeader from '@/src/components/EntityView/Header/Header';
 import { Toolset } from '@/src/models/dial/toolset';
 import { ApplicationRoute } from '@/src/types/routes';
@@ -20,7 +20,7 @@ const ToolsetProperties: FC<Props> = ({ names, selectedToolset, onChangeToolset 
     <div className="pt-3 w-full">
       <EntityHeader entity={selectedToolset} />
       <div className="flex-1 min-h-0 pt-4 flex flex-col">
-        <EntityMainProperties
+        <DeploymentProperties
           entity={selectedToolset}
           onChangeEntity={onChangeToolset}
           names={names}
