+++ apps/ai-dial-admin/src/server/entities/assets-api.ts
@@ -2,7 +2,7 @@ import { JWT } from 'next-auth/jwt';
 
 import { DEFAULT_ETAG, IF_MATCH, IF_NONE_MATCH } from '@/src/constants/api-headers';
 import { ROOT_FOLDER } from '@/src/constants/file';
-import { DeploymentAsset } from '@/src/models/dial/deployment-asset';
+import { Asset } from '@/src/models/dial/deployment-asset';
 import { DialFile } from '@/src/models/dial/file';
 import { DialPrompt } from '@/src/models/dial/prompt';
 import { ServerActionResponse } from '@/src/models/server-action';
@@ -47,14 +47,14 @@ export class AssetsApi extends BaseApi {
     token: JWT | null,
     path: string,
     type: ResourceType,
-  ): Promise<(DeploymentAsset | DialPrompt | DialFile)[] | null | undefined> {
+  ): Promise<(Asset | DialFile)[] | null | undefined> {
     const url = this.buildUrl(type, ResourceOperation.LIST);
     if (type === ResourceType.FILE) {
       return this.post(ResourceBasePaths[type], { path }, token).then((response) =>
         response === void 0 ? void 0 : (response as { items: DialFile[] })?.items || [],
       );
     } else {
-      const allItems: DeploymentAsset | DialPrompt[] = [];
+      const allItems: Asset[] = [];
       let nextToken: string | undefined = undefined;
 
       while (true) {
@@ -63,9 +63,7 @@ export class AssetsApi extends BaseApi {
           body.nextToken = nextToken;
         }
 
-        const response = (await this.post(url, body, token)) as
-          | { items: DeploymentAsset | DialPrompt[]; nextToken?: string }
-          | undefined;
+        const response = (await this.post(url, body, token)) as { items: Asset[]; nextToken?: string } | undefined;
 
         if (!response) break;
 
@@ -96,7 +94,7 @@ export class AssetsApi extends BaseApi {
 
   updateAssetWithEtag(
     token: JWT | null,
-    asset: DeploymentAsset | DialPrompt,
+    asset: Asset,
     type: ResourceType,
     etag: string,
   ): Promise<ServerActionResponse> {
@@ -104,20 +102,12 @@ export class AssetsApi extends BaseApi {
     return this.postAction(url, { ...asset }, token, { [IF_MATCH]: etag });
   }
 
-  updateAsset(
-    token: JWT | null,
-    asset: DeploymentAsset | DialPrompt,
-    type: ResourceType,
-  ): Promise<ServerActionResponse> {
+  updateAsset(token: JWT | null, asset: Asset, type: ResourceType): Promise<ServerActionResponse> {
     const url = this.buildUrl(type, ResourceOperation.UPDATE);
     return this.postAction(url, { ...asset }, token);
   }
 
-  createAsset(
-    asset: DeploymentAsset | DialPrompt,
-    type: ResourceType,
-    token: JWT | null,
-  ): Promise<ServerActionResponse> {
+  createAsset(asset: Asset, type: ResourceType, token: JWT | null): Promise<ServerActionResponse> {
     const url = this.buildUrl(type, ResourceOperation.CREATE);
     return this.postAction(url, { ...asset, folderId: asset.folderId || ROOT_FOLDER }, token);
   }
